+++ packages/form-data-parser/CHANGELOG.md
@@ -4,6 +4,28 @@ This is the changelog for [`form-data-parser`](https://github.com/mjackson/remix
 
 ## HEAD
 
+- BREAKING CHANGE: Override `parseFormData` signature so the upload handler is always last in the argument list. `parserOptions` are now an optional 2nd arg.
+
+```ts
+import { parseFormData } from '@mjackson/form-data-parser';
+
+// before
+await parseFormData(
+  request,
+  (fileUpload) => {
+    // ...
+  },
+  { maxFileSize },
+);
+
+// after
+await parseFormData(request, { maxFileSize }, (fileUpload) => {
+  // ...
+});
+```
+
+- Upgrade [`multipart-parser`](https://github.com/mjackson/remix-the-web/tree/main/packages/multipart-parser) to v0.8 to fix an issue where errors would crash the process when `maxFileSize` was exceeded (see #28)
+- Add an [example of how to use `form-data-parser`](https://github.com/mjackson/remix-the-web/tree/main/packages/form-data-parser/examples/node) together with [`file-storage`](https://github.com/mjackson/remix-the-web/tree/main/packages/file-storage) to handle multipart uploads on Node.js
 - Expand `FileUploadHandler` interface to support returning `Blob` from the upload handler, which is the superclass of `File`
 
 ## v0.6.0 (2025-01-15)